feat(session): v0.0.1
Implements session configuration

+++ application/entrypoint/src/main/kotlin/br/com/fairie/partypay/endpoints/menu/mapper/MenuDtoMapper.kt
@@ -37,5 +37,5 @@ fun Order.toDto(): OrderDTO =
     OrderDTO(
         name = name,
         description = description,
-        value = value.toDouble()
+        value = value
     )
